recreate the polygonArr object
